refacto: reorder objects api folder
