+++ ui/components/datasetItemWorkspace/src/lib/api/objectsApi/index.ts
@@ -5,6 +5,7 @@ License: CECILL-C
 -------------------------------------*/
 
 import type { Annotation, Entity, View } from "@pixano/core";
+import { PRE_ANNOTATION } from "../../constants";
 
 export type MView = Record<string, View | View[]>;
 
@@ -14,20 +15,22 @@ export const getObjectEntity = (ann: Annotation, entities: Entity[]): Entity | u
   return entities.find((entity) => entity.id === ann.data.entity_ref.id);
 };
 
-export * from "./getPixanoSource";
-export * from "./getTable";
-export * from "./getTopEntity";
-export * from "./mapObjectToKeypoints";
-export * from "./mapObjectWithNewStatus";
-export * from "./toggleObjectDisplayControl";
-export * from "./getTopEntity";
+export const getObjectsToPreAnnotate = (objects: Annotation[]): Annotation[] =>
+  objects.filter(
+    (object) => object.data.source_ref.name === PRE_ANNOTATION && !object.ui.review_state,
+  );
+
+export * from "./addOrUpdateSaveItem";
 export * from "./defineCreatedEntity";
 export * from "./defineCreatedObject";
 export * from "./defineObjectThumbnail";
+export * from "./sortAndFilterObjectsToAnnotate";
 export * from "./getPixanoSource";
+export * from "./getTable";
+export * from "./getTopEntity";
 export * from "./mapObjectToBBox";
 export * from "./mapObjectToKeypoints";
+export * from "./mapObjectToMasks";
 export * from "./mapObjectWithNewStatus";
 export * from "./toggleObjectDisplayControl";
 export * from "./updateExistingObject";
-export * from './mapObjectToMasks'
